Add release script for github
