+++ scripts/release/build.sbt
@@ -10,7 +10,4 @@ libraryDependencies += "io.circe" %% "circe-core" % "0.9.3"
 libraryDependencies += "io.circe" %% "circe-generic" % "0.9.3"
 libraryDependencies += "io.circe" %% "circe-parser" % "0.9.3"
 libraryDependencies += "com.github.scopt" %% "scopt" % "3.7.0"
-libraryDependencies += "org.http4s" %% "http4s-client" % "0.18.9"
-libraryDependencies += "org.http4s" %% "http4s-dsl" % "0.18.9"
-libraryDependencies += "org.http4s" %% "http4s-blaze-client" % "0.18.9"
-libraryDependencies += "org.http4s" %% "http4s-circe" % "0.18.9"
+libraryDependencies += "com.softwaremill.sttp" %% "core" % "1.1.12"
